Migrated zspell-next to zspell, zspell to zspell-old
