+++ Cargo.toml
@@ -4,7 +4,7 @@ members = [
     "crates/zspell",
     # "crates/zspell-py",
     "crates/zspell-cli",
-    "crates/zspell-next"
+    "crates/zspell-old"
 ]
 
 # Build with `cargo build --profile=release-debug`
