+++ Pipfile
@@ -8,6 +8,7 @@ maturin = "*"
 sphinx = "*"
 
 [dev-packages]
+black = "*"
 
 [requires]
 python_version = "3.10"
