--- crates/zspell-next/src/check.rs
@@ -1,1 +1,0 @@
-
