--- crates/zspell-next/src/meta.rs
@@ -1,1 +1,0 @@
-
