--- crates/zspell-next/src/suggestions.rs
@@ -1,1 +1,0 @@
-
