--- crates/zspell-next/src/system.rs
@@ -1,2 +1,0 @@
-#[cfg(test)]
-mod tests;
