--- crates/zspell-next/util/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "util"
-version = "0.2.2"
-edition = "2021"
-publish = false
-
-[dependencies]
-zspell = { path = "../" }
