+++ crates/zspell-old/src/affix/serde.rs
