+++ crates/zspell-old/src/dictionary.rs
