+++ crates/zspell-old/src/errors.rs
