+++ crates/zspell/Cargo.toml
@@ -43,11 +43,11 @@ unicode-segmentation = "1.10.0"
 [dev-dependencies]
 criterion = "0.4"
 predicates = "2.1"
+pretty_assertions = "1.3.0"
 tempfile = "3.3"
 util = { path = "util" }
 
 [build-dependencies]
-
 
 [[bench]]
 name = "datastructure"
