+++ crates/zspell/src/error.rs
