+++ crates/zspell/src/meta.rs
@@ -1,0 +1,1 @@
+//! Meta-related logic
