+++ crates/zspell/src/parser_affix/types.rs
