+++ crates/zspell/src/parser_affix/types_impl.rs
