--- crates/zspell/src/parser_dict.rs
@@ -1,1 +1,0 @@
-
