+++ crates/zspell/src/suggestions.rs
@@ -1,0 +1,1 @@
+//! Suggestion logic
