+++ crates/zspell/tests/affix_integration.rs
@@ -1,29 +1,29 @@
-use util::TestCollection;
-use zspell::Config;
+// use util::TestCollection;
+// use zspell::Config;
 
-#[test]
-fn affix_create_words() {
-    let mut afx = Config::new();
+// #[test]
+// fn affix_create_words() {
+//     let mut afx = Config::new();
 
-    let content = TestCollection::load("1_pfxsfx.test").afx_str;
+//     let content = TestCollection::load("1_pfxsfx.test").afx_str;
 
-    afx.load_from_str(content.as_str()).unwrap();
+//     afx.load_from_str(content.as_str()).unwrap();
 
-    assert_eq!(
-        afx.create_affixed_words("xxx", "A"),
-        vec!["xxx".to_string(), "aaxxx".to_string()]
-    );
-    assert_eq!(
-        afx.create_affixed_words("xxx", "B"),
-        vec!["xxx".to_string(), "xxxcc".to_string()]
-    );
-    assert_eq!(
-        afx.create_affixed_words("xxx", "AB"),
-        vec![
-            "xxx".to_string(),
-            "aaxxx".to_string(),
-            "xxxcc".to_string(),
-            "aaxxxcc".to_string()
-        ]
-    );
-}
+//     assert_eq!(
+//         afx.create_affixed_words("xxx", "A"),
+//         vec!["xxx".to_string(), "aaxxx".to_string()]
+//     );
+//     assert_eq!(
+//         afx.create_affixed_words("xxx", "B"),
+//         vec!["xxx".to_string(), "xxxcc".to_string()]
+//     );
+//     assert_eq!(
+//         afx.create_affixed_words("xxx", "AB"),
+//         vec![
+//             "xxx".to_string(),
+//             "aaxxx".to_string(),
+//             "xxxcc".to_string(),
+//             "aaxxxcc".to_string()
+//         ]
+//     );
+// }
